move tests to AbstractSingleThreadEventLoopTest so they are executed for other eventloop impls
