+++ transport-native-epoll/src/test/java/io/netty/channel/epoll/EpollEventLoopTest.java
@@ -41,6 +41,11 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class EpollEventLoopTest extends AbstractSingleThreadEventLoopTest {
 
     @Override
+    protected boolean supportsChannelIteration() {
+        return true;
+    }
+
+    @Override
     protected EventLoopGroup newEventLoopGroup() {
         return new EpollEventLoopGroup();
     }
